[+] Stop collaborating to a shared note

+++ utils/notes/delete.ts
@@ -4,6 +4,8 @@ import { createClient } from "@/utils/supabase/server";
 import { revalidateTag } from "next/cache";
 import { NoteResponse } from "@/utils/notes/save";
 import { DatabaseNote } from "@/types/note.types";
+import { auth, currentUser } from "@clerk/nextjs/server";
+import { DatabaseProfile } from "@/types/profiles.types";
 
 const deleteNote = async (id: DatabaseNote["id"]): Promise<NoteResponse> => {
   const supabase = createClient();
@@ -26,13 +28,25 @@ const deleteNote = async (id: DatabaseNote["id"]): Promise<NoteResponse> => {
   return { data: undefined, error: error ?? undefined };
 };
 
-const revokeSharing = async (noteId: DatabaseNote["id"], userId: string) => {
+const revokeSharing = async (
+  noteId: DatabaseNote["id"],
+  userId?: DatabaseProfile["user_id"],
+) => {
+  console.log("Revoking sharing: ", noteId, userId);
+
   const supabase = createClient();
+
+  const { userId: auth_uid } = auth();
+
+  if (!auth_uid) {
+    console.error("User not authenticated");
+    throw new Error("User not authenticated");
+  }
 
   const { data, error } = await supabase.from("user_notes").delete().eq(
     "note_id",
     noteId,
-  ).eq("user_id", userId)
+  ).eq("user_id", userId ?? auth_uid)
     .select();
 
   console.log("Revoked sharing: ", data, error);
@@ -40,14 +54,24 @@ const revokeSharing = async (noteId: DatabaseNote["id"], userId: string) => {
   revalidateTag("shared-with");
 
   if (error) {
+    console.error("Error revoking sharing: ", error);
     throw new Error(error.message);
   }
 
   if (!data.length) {
+    console.error("Cannot revoke sharing. Please try again later.");
     throw new Error("Cannot revoke sharing. Please try again later.");
   }
 
   return { data: undefined, error: error ?? undefined };
 };
 
-export { deleteNote, revokeSharing };
+const revokeCollaborating = async (
+  noteId: DatabaseNote["id"],
+) => {
+  console.log("Revoking collaboration: ", noteId);
+
+  return await revokeSharing(noteId);
+};
+
+export { deleteNote, revokeCollaborating, revokeSharing };
